refactor(deep-seek-harness): align module boundaries
