+++ src/main/services/deepSeekHarness/__tests__/config.test.ts
@@ -38,7 +38,7 @@ const {
   resolveDeepSeekHarnessEndpoint,
   rollbackDeepSeekHarnessConfig,
   writeDeepSeekHarnessConfig
-} = await import('../deepSeekHarnessConfig')
+} = await import('../config')
 
 const model = (partial: Partial<Model> = {}): Model =>
   ({
