+++ src/main/services/deepSeekHarness/config.ts
