+++ src/main/services/deepSeekHarness/index.ts
@@ -1,0 +1,1 @@
+export { DeepSeekHarnessService } from './DeepSeekHarnessService'
